backend/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
gunicorn==20.0.4
joblib==1.0.1
numpy==1.20.1
pandas==1.2.2
psycopg2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.23.2
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
